Rename model to models and correct imports
