+++ LibreCarSharing-frontend/src/lib/components/login/login_component.dart
@@ -3,7 +3,7 @@ import 'dart:html';
 
 import 'package:angular2/angular2.dart';
 
-import 'package:LibreCarSharingFrontend/model/user.dart';
+import 'package:LibreCarSharingFrontend/models/user.dart';
 
 @Component(
   selector: 'login',
