+++ LibreCarSharing-frontend/src/lib/models/car.dart
@@ -1,19 +1,20 @@
+import 'user.dart';
 import 'dart:convert';
-import 'package:LibreCarSharingFrontend/model/user.dart';
 
 class Car {
+  int id = 0;
   String name = "";
-  String typ = "";
+  String type = "";
   String location = "";
   String imageFile = "";
   bool status = true;
-  String information = "";
+  String info = "";
   User owner = null;
   int seats = 1;
   int color = 0x000000;
   String licencePlate = "";
 
-  Car(this.name, this.imageFile, this.typ, this.location, this.licencePlate);
+  Car(this.name, this.imageFile, this.type, this.location, this.licencePlate);
 
   String toJSON() {
     return JSON.encode({});
