+++ LibreCarSharing-frontend/src/lib/models/community.dart
