+++ LibreCarSharing-frontend/src/lib/models/user.dart
