+++ LibreCarSharing-frontend/src/pubspec.yaml
@@ -8,6 +8,7 @@ environment:
 dependencies:
   angular2: ^3.0.0
   angular_components: ^0.5.1
+  json_object: any
 
 dev_dependencies:
   angular_test: ^1.0.0-beta+2
